% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment_Metrics.R
\name{alignment_identity}
\alias{alignment_identity}
\title{Percent Identity of a Pairwise Alignment}
\usage{
alignment_identity(x)
}
\arguments{
\item{x}{An object of class \code{alignment} from \code{\link{align}}.}
}
\value{
A numeric percentage in \code{[0, 100]}.
}
\description{
Fraction of aligned columns (excluding gap columns) at which the
two sequences carry the same residue, expressed as a percentage.
}
\examples{
a <- align("ACGTGGA", "ACGTA")
alignment_identity(a)

}
