% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence_Utils.R
\name{codon_usage}
\alias{codon_usage}
\title{Codon Usage of a Coding Sequence}
\usage{
codon_usage(sequence)
}
\arguments{
\item{sequence}{A character string (a coding DNA/RNA sequence).}
}
\value{
A named integer vector of codon counts, sorted decreasing.
}
\description{
Splits a sequence into non-overlapping codons (triplets) and
tabulates their frequencies. Trailing bases that do not form a full codon
are ignored.
}
\examples{
codon_usage("ATGAAATAG")

}
