% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{conservation_scores}
\alias{conservation_scores}
\title{Per-column Conservation Scores of an MSA}
\usage{
conservation_scores(x)
}
\arguments{
\item{x}{An object of class \code{msa}.}
}
\value{
A numeric vector, one conservation score per column.
}
\description{
Computes, for each alignment column, the fraction of rows that
carry the most common residue in that column (gaps excluded). A value of
1 means fully conserved.
}
\examples{
conservation_scores(msa_align(c(a = "ACGT", b = "ACGA", c = "ACGT")))

}
