% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq_Manipulation.R
\name{DNASequence}
\alias{DNASequence}
\title{Create a DNASequence Object}
\usage{
DNASequence(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence. Must contain only A, C, G, and T.}
}
\value{
An object of class \code{DNASequence} with the sequence attribute.
}
\description{
Constructs an S3 object of class \code{DNASequence} representing a DNA sequence.
This constructor validates the input to ensure it contains only valid DNA bases (A, C, G, T).
}
\examples{
dna_seq <- DNASequence("ATGCGC")
print(dna_seq)

}
