% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq_Manipulation.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC Content of a DNA Sequence}
\usage{
gc_content(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence. Must contain only A, C, G, and T.}
}
\value{
A numeric value representing the GC content percentage of the DNA sequence.
}
\description{
This function calculates the GC content (percentage of guanine and cytosine) in a given DNA sequence.
}
\examples{
gc_content("ATGC") # Returns 50
gc_content("GCGC") # Returns 100
gc_content("ATAT") # Returns 0

}
