% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq_Manipulation.R
\name{gc_content.S3}
\alias{gc_content.S3}
\title{Calculate GC Content of a DNA Sequence (S3 Method)}
\usage{
gc_content.S3(sequence)
}
\arguments{
\item{sequence}{An object of class \code{DNASequence}.}
}
\value{
A numeric value representing the GC content percentage of the DNA sequence.
}
\description{
Calculates the GC content (percentage of guanine and cytosine) in a DNA sequence represented by a \code{DNASequence} object.
}
\examples{
dna_seq <- DNASequence("ATGCGC")
gc_content.S3(dna_seq)  # Returns 66.67

}
