% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence_Utils.R
\name{kmer_count}
\alias{kmer_count}
\title{Count k-mers in a Sequence}
\usage{
kmer_count(sequence, k = 3)
}
\arguments{
\item{sequence}{A character string.}

\item{k}{k-mer length (positive integer).}
}
\value{
A named integer vector of k-mer counts, sorted in decreasing order.
}
\description{
Counts all overlapping substrings of length \code{k}.
}
\examples{
kmer_count("ATATAT", 2)  # AT = 3, TA = 2

}
