% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylo.R
\name{msa_distance}
\alias{msa_distance}
\title{Distance Matrix from a Multiple Sequence Alignment}
\usage{
msa_distance(x)
}
\arguments{
\item{x}{An object of class \code{msa}.}
}
\value{
A symmetric \code{dist}-compatible numeric matrix of pairwise
distances, with sequence names as row/column labels.
}
\description{
Computes a pairwise distance matrix from an MSA using
\emph{p-distance} (proportion of differing, non-gap columns between each
pair of sequences). This is the standard hand-off point to phylogenetics.
}
\examples{
m <- msa_align(c(a = "ACGTGGAA", b = "ACGTGCAA", c = "TTGTGGATA"))
msa_distance(m)

}
