% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylo.R
\name{nj_tree}
\alias{nj_tree}
\title{Build and Plot a Neighbour-Joining Tree from an MSA}
\usage{
nj_tree(x, plot = TRUE)
}
\arguments{
\item{x}{An object of class \code{msa}.}

\item{plot}{Logical; if \code{TRUE} (default) the tree is plotted.}
}
\value{
An object of class \code{phylo} (from \pkg{ape}), invisibly.
}
\description{
Computes a distance matrix from an MSA and builds a
neighbour-joining phylogenetic tree using the \pkg{ape} package. This
bridges alignment straight into phylogenetics in one call.
}
\examples{
\donttest{
m <- msa_align(c(a = "ACGTGGAA", b = "ACGTGCAA", c = "TTGTGGATA", d = "ACGTGGTA"))
nj_tree(m)
}

}
