% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA_Visualization.R
\name{plot.alignment}
\alias{plot.alignment}
\title{Plot a Pairwise Alignment}
\usage{
\method{plot}{alignment}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alignment}.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
ggplot2 tile view of a pairwise alignment, colouring matches,
mismatches, and gaps so regions of similarity stand out.
}
\examples{
\donttest{
plot(align("ACGTGGA", "ACGTGCA"))
}

}
