% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA_Visualization.R
\name{plot.msa}
\alias{plot.msa}
\title{Plot a Multiple Sequence Alignment}
\usage{
\method{plot}{msa}(x, color_by = c("residue", "conservation"), ...)
}
\arguments{
\item{x}{An object of class \code{msa}.}

\item{color_by}{\code{"residue"} (default) colours each nucleotide/amino
acid; \code{"conservation"} shades columns by how conserved they are.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
ggplot2 tile view of an MSA, colouring each residue and
optionally shading columns by conservation.
}
\examples{
\donttest{
m <- msa_align(c(a = "ACGTGGAA", b = "ACGTGCAA", c = "ACGTGGATA"))
plot(m)
}

}
