% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{print.msa}
\alias{print.msa}
\title{Print an MSA}
\usage{
\method{print}{msa}(x, width = 60, ...)
}
\arguments{
\item{x}{An object of class \code{msa}.}

\item{width}{Columns per printed block.}

\item{...}{Ignored.}
}
\value{
The msa object, invisibly.
}
\description{
Prints the aligned rows plus a consensus line.
}
