% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq_Manipulation.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse Complement of a DNA Sequence}
\usage{
reverse_complement(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence. Must contain only A, C, G, and T.}
}
\value{
A character string representing the reverse complement of the input DNA sequence.
}
\description{
This function computes the reverse complement of a given DNA sequence by reversing it and substituting complementary bases.
}
\examples{
sequence <- ("ATCGCTCA") # Returns "TGAGCGAT"

}
