% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq_Manipulation.R
\name{transcribe_dna}
\alias{transcribe_dna}
\title{Transcribe DNA to RNA}
\usage{
transcribe_dna(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA sequence. Must contain only A, C, G, and T.}
}
\value{
A character string representing the transcribed RNA sequence.
}
\description{
This function converts a DNA sequence into an RNA sequence by replacing thymine (T) with uracil (U).
}
\examples{
transcribe_dna("ATGC") # Returns "AUGC"
transcribe_dna("TTAA") # Returns "UUAA"

}
