% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_motif.R
\name{find_motif}
\alias{find_motif}
\title{Locates all positions of a motif in a sequence.}
\usage{
find_motif(seq, motif)
}
\arguments{
\item{seq}{DNA sequence (character string)}

\item{motif}{Motif to search for (character string)}
}
\value{
#A vector of starting positions where the motif occurs
}
\description{
Locates all positions of a motif in a sequence.
}
